+++ core_gearworks_lib/source/lib/alias.h
@@ -1,0 +1,13 @@
+#ifndef ALIAS_H
+#define ALIAS_H
+
+#include <gearworks.h>
+
+typedef std::string string;
+
+typedef glm::vec2 vec2;
+typedef glm::vec3 vec3;
+typedef glm::vec4 vec4;
+typedef glm::mat4 mat4;
+
+#endif // Header guard